feat(media-modal): add flow and modal coverage
